Move themes to root
